Extract interface to allow more metrics enums.
